Week 4 Update (See extended description)
+ Added clearFinishedTasks(); which can be fired from the console.

+ Added addTask(taskName, dueDate, taskComplete) that can be fired from
the console.
+ taskName is a string
+ dueDate is a string in the format of yyyy, mm, dd
+ taskComplete is a boolean
+ addTask fires off a small set of functions that refresh the list with
the newest task at the bottom

+ printTask() can be fired on it's own from the console, if you really
wanted to duplicate your list for no reason. Maybe you're just a
masochist.

+ uses strict mode now, if only to get JSHint to shut up. Several of the
things it complained about "===" instead of "==" and my use of
semicolons at the end of functions were really semantic stuff. I fixed
the "===" so that strict mode worked better, but I left my semicolons
because they are merely a dialect choice.
+Use of strict mode requires defining throwaway variables for running
for loops, which seems silly and adds more work.

+ Space Marine Gauss Rifle now does + 4 damage per shot.